Daily-tracker/requirements.txt
streamlit
pandas
google-generativeai